requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
django-filter